pyproject.toml
[project]
name = "prebid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "dotenv>=0.9.9",
    "openai>=2.9.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "ruff>=0.14.8",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.45",
]
